atualização da classe Tipos Variavel

+++ ProjetoTiposVariavel/SalarioMensal.java
@@ -1,8 +1,8 @@
 public class SalarioMensal {
 	public static void main(String[] args) {
 	
-  Float salarioMensal = 2300.61F ;    // Salario Base
-	salarioMensal = 2300.61F + 1500.00F; // Hora Extras
+  Double salarioMensal = 2300.61 ;    // Salario Base
+	salarioMensal = 2300.61 + 1500.00; // Hora Extras
 	
 	System.out.println("Salario Mensal: R$ " + salarioMensal);
